requirements.txt
# Core pipeline
openai-whisper>=20231117
transformers>=4.45.0
qwen-vl-utils>=0.0.8
torch>=2.0.0
torchvision>=0.15.0

# Video processing
opencv-python>=4.8.0
ffmpeg-python>=0.2.0

# Utilities
accelerate>=0.26.0       # For device_map="auto" in Qwen2-VL
Pillow>=10.0.0
numpy>=1.24.0